Conflicts:
- .travis.yml (changes made to this file moved into dev/travis/script.sh)
- dev/tests/integration/IntegationTestsForTravis.sh (only incoming change was the copyright date change)
- dev/tests/integration/phpunit.xml.dist

+++ app/code/Magento/Bundle/Model/Product/SaveHandler.php
@@ -1,20 +1,20 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © 2016 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 
 namespace Magento\Bundle\Model\Product;
 
 use Magento\Bundle\Api\ProductOptionRepositoryInterface as OptionRepository;
-use Magento\Framework\Model\Entity\MetadataPool;
-use Magento\Framework\Model\ResourceModel\Db\ProcessEntityRelationInterface;
+use Magento\Framework\EntityManager\MetadataPool;
 use Magento\Bundle\Api\ProductLinkManagementInterface;
+use Magento\Framework\EntityManager\Operation\ExtensionInterface;
 
 /**
  * Class SaveHandler
  */
-class SaveHandler
+class SaveHandler implements ExtensionInterface
 {
     /**
      * @var MetadataPool
@@ -49,25 +49,36 @@ class SaveHandler
     /**
      * @param string $entityType
      * @param object $entity
-     * @return object
+     * @param array $arguments
+     * @return \Magento\Catalog\Api\Data\ProductInterface|object
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function execute($entityType, $entity)
+    public function execute($entityType, $entity, $arguments = [])
     {
         $bundleProductOptions = $entity->getExtensionAttributes()->getBundleProductOptions();
         if ($entity->getTypeId() !== 'bundle' || empty($bundleProductOptions)) {
             return $entity;
         }
-        /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
-        foreach ($this->optionRepository->getList($entity->getSku()) as $option) {
-            $this->removeOptionLinks($entity->getSku(), $option);
-            $this->optionRepository->delete($option);
-        }
 
-        $options = $entity->getExtensionAttributes()->getBundleProductOptions() ?: [];
-        foreach ($options as $option) {
-            $option->setOptionId(null);
-            $this->optionRepository->save($entity, $option);
+        if (!$entity->getCopyFromView()) {
+            /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
+            foreach ($this->optionRepository->getList($entity->getSku()) as $option) {
+                $this->removeOptionLinks($entity->getSku(), $option);
+                $this->optionRepository->delete($option);
+            }
+
+            $options = $bundleProductOptions ?: [];
+            foreach ($options as $option) {
+                $option->setOptionId(null);
+                $this->optionRepository->save($entity, $option);
+            }
+        } else {
+            //save only labels and not selections + product links
+            $options = $bundleProductOptions ?: [];
+            foreach ($options as $option) {
+                $this->optionRepository->save($entity, $option);
+                $entity->setCopyFromView(false);
+            }
         }
         return $entity;
     }
